move EitherUtils into util package
