+++ app/uk/gov/hmrc/hec/testonly/models/SaveTaxCheckRequest.scala
@@ -20,7 +20,7 @@ import play.api.libs.json.{Json, Reads}
 import uk.gov.hmrc.hec.models.ids.{CRN, GGCredId}
 import uk.gov.hmrc.hec.models.licence.LicenceType
 import uk.gov.hmrc.hec.models.{CorrectiveAction, DateOfBirth, TaxYear}
-import uk.gov.hmrc.hec.models.EitherUtils.eitherFormat
+import uk.gov.hmrc.hec.util.EitherUtils.eitherFormat
 import uk.gov.hmrc.hec.models.hecTaxCheck.{HECTaxCheckCode, HECTaxCheckSource}
 
 import java.time.format.DateTimeFormatter
